requirements-backend.txt
# FastAPI Backend Dependencies for PyGliderCG
# Python 3.8+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and settings management
pydantic==2.4.2
pydantic-settings==2.0.3

# Environment variables
python-dotenv==1.2.2

# Database
duckdb==1.3.2

# Authentication and Security
PyJWT==2.12.0
bcrypt==4.3.0

# Request parsing
python-multipart==0.0.6

# Production server (optional, use with: gunicorn -w 4 -k uvicorn.workers.UvicornWorker backend.main:app)
gunicorn==21.2.0

# Monitoring and logging (optional)
# python-json-logger==2.0.7

